requirements.txt
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.4.0
plotly==5.13.0
seaborn==0.12.2
streamlit==1.17.0
